+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-23T17:36:36.357Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-24T03:24:22.071Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")