+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVTRX9PYXM3W5EHJY7T5KGE9"
+        "01KVVTHHQTTPRSV38AGP9EFVMQ"
     )
     r"""The unique identifier of the data source"""