+++ packages/orq-rc/RELEASES.md
@@ -8939,3 +8939,13 @@ Based on:
 - [python v4.11.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.43 - packages/orq-rc
+
+## 2026-06-24 03:26:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.44 - packages/orq-rc